Enbaled keyboard control
